+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## History
-* 2021/04/08 ver2.25.11
+* 2021/04/** ver2.25.12
+	* Killer cage submode in Cage mode. Automatically drawing killer cages.
+	* Fixed bug in multisolution feature for duplicate function.
+	* Fixed bug in Custom Color for Undo.
+* 2021/04/12 ver2.25.11
 	* Cookies Feature - User can now save Grid Type, Tab Settings, Theme and Reload button status locally by clicking Save Settings: YES in the Settings Button. When penpa loads, it will use the User preferences. 
 	* Number shortcuts to switch between different styles for Surface mode.
 	* Improvement to multisolution feature to detect changes in the grid.